vpr_bot/requirements.txt
aiogram==3.3.0
openai>=1.0.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
